requirements.txt
beautifulsoup4
requests
pillow
fpdf
PyPDF2
